Migrate kernels into dbk_kernels
